Read bundle registry contract address from env
Similar to other config parameters WEB3_NODEPRIVATEKEY, MONGODB_URI,
WEB3_RPC I would like to be able to provide this parameter using the
environment.

The use case is that it simplifies deployments if as much as possible
can be injected using environment variables rather than relying on files
and Docker data volumes.

https://12factor.net/config