Read git commit info from env (#175)
* Read git commit info from env

We run ambnode as a container without git and we don't copy the git
directory to the container. The /nodeInfo endpoint fails for this
reason because it tries to shell out to git and the executable does not
exist in the container.

Instead of adding the git history and installing git in the container
(and shelling out for each request) we can read the git commit as an
environment variable.

At build time of the container we set the build arg to the git commit
and it persists when the container is running.

E.g
```
$ docker build --build-arg GIT_COMMIT=$(git rev-parse --short HEAD) \
	-t ambrosus/ambrosus-node .
```

I also add a label for the git commit so that we can track what version
is running using our monitoring tools.

I've shortened the git commit from the longer version to the commit-ish
which is 7 characters, but this is up for debate and I can easily change
it back.