Refactor config (#178)
An attempt to unify the config which is scattered all over the place
into a single file. Configuration variables are read from the
environment as recommended in https://12factor.net/config.

Once configuration is read, it cannot be altered as the object is frozen.

For test we specify variables in ./test/mocha_env.js which is read by
./test/mocha.opts on startup. Most tests use the default mocha config
options, but it's possible to change it on a test-by-test basis by
providing custom config options to the apparatus. See
https://github.com/ambrosus/ambrosus-node/blob/refactor-config/test/helpers/apparatus.js#L25-L30.

For staging / prod they the variables are injected into the environment
by the deployment tooling.

For local setup I'd recommend using docker-compose if the process is ran
as a container or using .envrc if it's ran with Node.js on the host.