Fix: incremental backoff when failing to upload bundles
Parity RPC servers may be rate limited and sending too HTTP requests
causes the HTTP client to be restricted with 429 errors. In order to
safely upload a bundle we implement incremental back off mechanism which
starts at 1s and is multiplied by 2 for each consecutive HTTP request
that hits the rate limit.