Add process instrumentation (#324)
I've seen unhandledRejection errors bubble up to the process when
operating Hermes workers in production. This causes a zombie process
which isn't operational, but continues running.

The commit introduces the following changes:

on unhandledRejection or uncaughtException errors that are not handled
and bubble up to the process, do the following:

- exit with a status code of 1 so that the process is restarted by it's
  supervisor in a fresh state (in our case by the Docker daemon)

- log the error

- report the error to Sentry (if Sentry is enabled)

Additionally we enable Sentry for the Hermes and Atlas workers rather
than having it run only on the server.